+++ CHANGELOG.md
@@ -1,5 +1,35 @@
 # Changelog
 
+## [0.5.0](https://github.com/nikrich/hungry-ghost-hive/compare/v0.4.1...v0.5.0) (2026-02-06)
+
+
+### Features
+
+* make QA agent scaling configurable via hive.config.yaml ([49471e1](https://github.com/nikrich/hungry-ghost-hive/commit/49471e14b4fcf03f5d785a97b898dbcbcfe798a4))
+
+## [0.4.1](https://github.com/nikrich/hungry-ghost-hive/compare/v0.4.0...v0.4.1) (2026-02-06)
+
+
+### Bug Fixes
+
+* optimize manager daemon to eliminate N+1 queries ([64744fc](https://github.com/nikrich/hungry-ghost-hive/commit/64744fc04b381789351c177f579a065231fde823))
+
+## [0.4.0](https://github.com/nikrich/hungry-ghost-hive/compare/v0.3.1...v0.4.0) (2026-02-06)
+
+
+### Features
+
+* add comprehensive cleanup command for orphaned resources ([7013aa2](https://github.com/nikrich/hungry-ghost-hive/commit/7013aa29d8c45da023f4fa389e59736b3bb25e35))
+* **db:** add database indexes for query performance ([2fc3b06](https://github.com/nikrich/hungry-ghost-hive/commit/2fc3b060559508f9b2ba87a5742a760dc7f9ccb3))
+* **scheduler:** refactor duplicated agent spawn methods into generic spawnAgent ([049eae2](https://github.com/nikrich/hungry-ghost-hive/commit/049eae2f475febeb5f386f2ba2e6defd99f0f52f))
+* update default model IDs to latest Claude versions ([a551fca](https://github.com/nikrich/hungry-ghost-hive/commit/a551fca2edbbf5f065dcc522fe3efd05670209ce))
+
+
+### Bug Fixes
+
+* change versionMap from let to const to fix linting error ([85f409b](https://github.com/nikrich/hungry-ghost-hive/commit/85f409bd4e4878d6e17dff3f32e0c41bfb89a0ae))
+* change versionMap from let to const to fix linting error ([a52e0d0](https://github.com/nikrich/hungry-ghost-hive/commit/a52e0d0b2819a7dac72e8df6a73702ff9c70a10a))
+
 ## [0.3.1](https://github.com/nikrich/hungry-ghost-hive/compare/v0.3.0...v0.3.1) (2026-02-06)